requirements.txt
alembic==1.14.1
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.4
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.41.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
feature-engine==1.8.2
Flask==3.1.0
fonttools==4.55.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
hyperopt==0.2.7
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mlflow==2.19.0
mlflow-skinny==2.19.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==14.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
shap==0.46.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
